[tests] reorganize tests
